+++ src/app/layout/ads/ads.component.ts
@@ -4,6 +4,7 @@ import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { CommonServiceService } from '../../common-service.service';
+import { ConfirmationDialogService } from '../components/confirmation-dialog/confirmation-dialog.service';
 export interface DialogData {
   data1: any;
   ind: any;
@@ -20,7 +21,8 @@ export interface DialogData {
 export class AdsComponent implements OnInit {
   constructor(public dialog: MatDialog,
     private snackBar: MatSnackBar,
-    private cds: CommonServiceService) { }
+    private cds: CommonServiceService,
+    private cnfr: ConfirmationDialogService) { }
   visible: any;
   AdsCollection: any[];
   OfferCollection: any[];
@@ -33,7 +35,6 @@ export class AdsComponent implements OnInit {
   nextProcess() {
     this.visible = true;
     this.cds.getAllAds(this.cds.tokenLogin).subscribe(response => {
-
       this.AdsCollection = this.getAds(response["ads"]);
       this.cds.getAllOffer(this.cds.tokenLogin).subscribe(response => {
         this.visible = false;
@@ -112,16 +113,25 @@ export class AdsComponent implements OnInit {
     });
   }
   deleteAds(val) {
-    this.cds.deleteAds(val.adsId, this.cds.tokenLogin).subscribe(response => {
-      this.snackBar.open((response ? response["message"] : ""), "", {
-        duration: 2000,
+    this.cnfr.confirm('Please confirm..', 'Do you really want to ... ?')
+      .then((confirmed) => {
+        console.log('User confirmed:', confirmed);
+        if (confirmed) {
+          this.cds.deleteAds(val.adsId, this.cds.tokenLogin).subscribe(response => {
+            this.snackBar.open((response ? response["message"] : ""), "", {
+              duration: 2000,
+            });
+            this.nextProcess();
+          }, error => {
+            this.snackBar.open((error.error.error ? error.error.error.message : error.error.message), "", {
+              duration: 2000,
+            });
+          });
+        }
+      })
+      .catch(() => {
+        console.log('User dismissed the')
       });
-      this.nextProcess();
-    }, error => {
-      this.snackBar.open((error.error.error ? error.error.error.message : error.error.message), "", {
-        duration: 2000,
-      });
-    });
   }
 }
