+++ src/app/layout/dispenser/dispenser.component.ts
@@ -4,6 +4,7 @@ import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { FormGroup, FormBuilder, Validators } from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { CommonServiceService } from '../../common-service.service';
+import { ConfirmationDialogService } from '../components/confirmation-dialog/confirmation-dialog.service';
 export interface DialogData {
   data1: any;
   ind: any;
@@ -20,7 +21,8 @@ export class DispenserComponent implements OnInit {
   public DispenserCollection: any[];
   constructor(public dialog: MatDialog,
     private snackBar: MatSnackBar,
-    private cds: CommonServiceService) { }
+    private cds: CommonServiceService,
+    private cnfr: ConfirmationDialogService) { }
 
   ngOnInit() {
     if (this.cds.tokenLogin === undefined) {
@@ -93,16 +95,25 @@ export class DispenserComponent implements OnInit {
   }
 
   deleteDispenser(val) {
-    this.cds.deleteDispenser(val.id, this.cds.tokenLogin).subscribe(response => {
-      this.snackBar.open((response ? response["message"] : ""), "", {
-        duration: 2000,
+    this.cnfr.confirm('Please confirm..', 'Do you really want to ... ?')
+      .then((confirmed) => {
+        console.log('User confirmed:', confirmed);
+        if (confirmed) {
+          this.cds.deleteDispenser(val.id, this.cds.tokenLogin).subscribe(response => {
+            this.snackBar.open((response ? response["message"] : ""), "", {
+              duration: 2000,
+            });
+            this.nextProcess();
+          }, error => {
+            this.snackBar.open((error.error.error ? error.error.error.message : error.error.message), "", {
+              duration: 2000,
+            });
+          });
+        }
+      })
+      .catch(() => {
+        console.log('User dismissed the')
       });
-      this.nextProcess();
-    }, error => {
-      this.snackBar.open((error.error.error ? error.error.error.message : error.error.message), "", {
-        duration: 2000,
-      });
-    });
   }
 }
